refactor(driver): extract FlushPlan typestate to driver/flush_plan.rs (refactor 2.1d)
Fourth commit of the Phase 2.1 mechanical split. Moves the Phase 1.7
typestate machinery (`FlushPlan<S>` + the sealed `Sendable` trait +
`NoEntries`/`Collected`/`HoldArmed` phantom markers + `SendResult` +
`RollbackPackage`) out of driver/mod.rs into a dedicated sibling module.

- `crates/based-rollup/src/driver/flush_plan.rs` (new, 321 lines):
  - Full module block comment explaining the typestate rationale.
  - Sealed `flush_plan_sealed` submodule with `Sealed` super-trait.
  - Phantom markers `NoEntries`, `Collected`, `HoldArmed`.
  - `Sendable` trait (sealed) + impls for `NoEntries` / `HoldArmed`.
  - `SendResult` (`#[must_use]`) + `RollbackPackage`.
  - `FlushPlan<S>` struct + all impl blocks:
    - `FlushPlan<NoEntries>::new_blocks_only`
    - `FlushPlan<Collected>::new_collected`, `::arm_hold`
    - `FlushPlan<S: Sendable>::submit_via`
    - `FlushPlan<S>::{block_count, entry_count, is_empty}`
  - Imports: `super::pending_queue::PendingL1SubmissionQueue`,
    `super::hold::EntryVerificationHold`,
    `crate::cross_chain::CrossChainExecutionEntry`,
    `crate::proposer::{GasPriceHint, PendingBlock, Proposer}`,
    `alloy_primitives::B256`, `std::marker::PhantomData`.

- `crates/based-rollup/src/driver/mod.rs` (-303 lines → 4824 total):
  - `mod flush_plan;` declaration.
  - `pub use flush_plan::{Collected, FlushPlan, HoldArmed, NoEntries,
    RollbackPackage, SendResult, Sendable};` — re-export keeps the
    `use super::*` pattern in driver tests working.
  - `use std::marker::PhantomData;` removed from the top-level imports
    (no remaining direct use in mod.rs).

Cumulative extraction progress:
- mod.rs: 5509 → 4824 lines (-685 / -12.4%)
- 3 new sibling modules totaling 722 lines of well-bounded types.

`cargo check`, `cargo clippy --all-features -- -D warnings`, and
`cargo nextest -p based-rollup` (608/608) all pass. Zero behavioral
change.

Next: 2.1e extracts `TxJournalEntry` + the tx-journal impl methods
(`journal_block_transactions`, `save_tx_journal`, `load_tx_journal`,
`prune_tx_journal`) to `driver/journal.rs`.
